backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-sqlalchemy==2.5.1
flask-jwt-extended==4.3.1
python-dotenv==0.19.0
psycopg2-binary==2.9.1
werkzeug==2.0.1
redis==4.5.1
